// src/interfaces/ISoundtrack.ts
export interface ISoundtrack {
    id: number;             // Unique identifier for the soundtrack
    title: string;          // Title of the soundtrack
    description: string;    // Description of the soundtrack
    url: string;            // URL to the soundtrack file
    thumbnail: string;      // URL to the thumbnail image
}